+++ pages/_app.tsx
@@ -4,10 +4,10 @@ import { ThemeProvider } from '@mui/material/styles';
 import { NextIntlProvider } from 'next-intl';
 import type { AppProps } from 'next/app';
 import Head from 'next/head';
+import { wrapper } from '../api/index';
 import Footer from '../components/Footer';
 import TopBar from '../components/TopBar';
 import createEmotionCache from '../config/emotionCache';
-import { wrapper } from '../store/index';
 import '../styles/globals.css';
 import theme from '../styles/theme';
