+++ pages/therapy-booking.tsx
@@ -8,6 +8,7 @@ import { useTranslations } from 'next-intl';
 import Image from 'next/image';
 import Script from 'next/script';
 import { useState } from 'react';
+import { useAppSelector } from '../common/hooks';
 import { therapyFaqs } from '../common/therapyFaqs';
 import Faqs from '../components/Faqs';
 import Header from '../components/Header';
@@ -18,6 +19,7 @@ import illustrationConfidential from '../public/illustration_confidential.svg';
 import illustrationDateSelector from '../public/illustration_date_selector.svg';
 import illustrationLeafMix from '../public/illustration_leaf_mix.svg';
 import illustrationTeaPeach from '../public/illustration_tea_peach.png';
+import { AppState } from '../store';
 import { rowStyle } from '../styles/common';
 
 const steps: Array<ImageTextItem> = [
@@ -47,6 +49,10 @@ const TherapyBooking: NextPage = () => {
   const t = useTranslations('TherapyBooking');
   const tS = useTranslations('Shared');
   const [widgetOpen, setWidgetOpen] = useState(false);
+
+  const { user } = useAppSelector((state: AppState) => state);
+
+  console.log('user', user);
 
   const widgetConfig = {
     widget_type: 'iframe',
